Move DAC interaction code to a separate module
